# api/openapi.yaml
openapi: 3.0.3
info:
  description: |
    API for a mobile safety application that allows users to post and subscribe to local safety concerns.
  title: EPA (Event Posting App) API
  version: 1.0.0
servers:
- url: /
tags:
- description: Service health and metadata.
  name: System
- description: Identity management including native and social OAuth2 exchanges.
  name: Authentication
paths:
  /v1/status:
    get:
      operationId: get_api_status
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
          description: API is healthy
      summary: Check API health
      tags:
      - System
  /v1/auth/register:
    post:
      description: Creates a user account. Users must be authenticated via email to
        avoid spam.
      operationId: register_new_user
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCreated"
          description: User created successfully
      summary: Register a new user
      tags:
      - Authentication
  /v1/auth/login:
    post:
      description: Returns a JWT for application access.
      operationId: login_with_password
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
          description: Successful authentication returns access and a session token.
      summary: Login with email/password
      tags:
      - Authentication
  /v1/auth/social/google:
    post:
      description: Exchanges a Google ID Token for an EPA-issued JWT.
      operationId: authenticate_with_google
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SocialTokenExchange"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
          description: Successful authentication returns access and session token.
      summary: Google OAuth2 Exchange
      tags:
      - Authentication
  /v1/auth/social/apple:
    post:
      description: Exchanges an Apple Identity Token for an EPA-issued JWT.
      operationId: authenticate_with_apple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppleTokenExchange"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
          description: Successful authentication returns access and session token.
      summary: Apple Sign-In Exchange
      tags:
      - Authentication
  /v1/auth/session:
    post:
      description: Uses a session token to generate a new short-lived access JWT.
      operationId: renew_session_token
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
          description: Successful authentication returns access and a session token.
      security:
      - BearerAuth: []
      summary: Session Token Renewal
      tags:
      - Authentication
components:
  responses:
    TokenResponse:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AuthToken"
      description: Successful authentication returns access and a session token.
  schemas:
    Status:
      example:
        status: healthy
        version: 1.0.0
      properties:
        status:
          type: string
        version:
          type: string
      title: Status
      type: object
    UserRegistration:
      example:
        password: password
        email: email
        username: username
      properties:
        email:
          format: email
          title: email
          type: string
        password:
          format: password
          title: password
          type: string
        username:
          title: username
          type: string
      required:
      - email
      - password
      - username
      title: UserRegistration
      type: object
    LoginRequest:
      example:
        password: password
        email: email
      properties:
        email:
          title: email
          type: string
        password:
          format: password
          title: password
          type: string
      required:
      - email
      - password
      title: LoginRequest
      type: object
    SocialTokenExchange:
      example:
        id_token: id_token
      properties:
        id_token:
          description: Token from social provider SDK.
          title: id_token
          type: string
      required:
      - id_token
      title: SocialTokenExchange
      type: object
    AppleTokenExchange:
      example:
        identity_token: identity_token
        full_name: full_name
      properties:
        identity_token:
          title: identity_token
          type: string
        full_name:
          title: full_name
          type: string
      required:
      - identityToken
      title: AppleTokenExchange
      type: object
    UserCreated:
      example:
        user_id: some_user_id
      properties:
        user_id:
          title: user_id
          type: string
      title: UserCreated
      type: object
    AuthToken:
      example:
        access_token: access_token
        session_token: session_token
        token_type: Bearer
        expires_in: 3600
      properties:
        access_token:
          title: access_token
          type: string
        session_token:
          title: session_token
          type: string
        token_type:
          example: Bearer
          title: token_type
          type: string
        expires_in:
          example: 3600
          title: expires_in
          type: integer
      title: AuthToken
      type: object
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      description: |
        Enter your JWT access token. This allows the RESTful API to verify your identity before allowing posts or subscriptions.
      scheme: bearer
      type: http